use message dispatching instead of busy-waiting
As cow-notify has no special requirements for a mainloop we can
avoid doing sleep cycles. This is not only saving a few cpu cycles,
but gives us also 'instant' notifications instead of ~1 sec delays.
.
The code also passes the used DBusConnection around instead of
storing it in a global variable - that's again more a taste than a
really needed change.